% Generated by roxygen2 (4.0.2): do not edit by hand
\name{agg.pval}
\alias{agg.pval}
\title{P-value aggregation (Meinshausen et al 2009)}
\usage{
agg.pval(gamma, pval)
}
\arguments{
\item{gamma}{see Meinshausen et al 2009}

\item{pval}{vector of p-values}
}
\value{
inf-quantile aggregated p-value
}
\description{
P-value aggregation
}
\author{
n.stadler
}
\keyword{internal}

