% Generated by roxygen2 (4.0.2): do not edit by hand
\name{agg.score.iriz.scale}
\alias{agg.score.iriz.scale}
\title{Irizarry aggregate score (scale)}
\usage{
agg.score.iriz.scale(ttstat, geneset, gene.name)
}
\arguments{
\item{ttstat}{no descr}

\item{geneset}{no descr}

\item{gene.name}{no descr}
}
\value{
no descr
}
\description{
Irizarry aggregate score (scale)
}
\author{
n.stadler
}
\keyword{internal}

