% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggpval}
\alias{aggpval}
\title{Meinshausen p-value aggregation}
\usage{
aggpval(pval, gamma.min = 0.05)
}
\arguments{
\item{pval}{Vector of p-values.}

\item{gamma.min}{See inf-quantile formula of Meinshausen et al 2009 (default=0.05).}
}
\value{
Aggregated p-value.
}
\description{
Meinshausen p-value aggregation.
}
\details{
Inf-quantile formula for p-value aggregation presented in Meinshausen et al 2009.
}
\examples{
pval=runif(50)
aggpval(pval)
}
\author{
n.stadler
}

