% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildDotPlotDataFrame}
\alias{buildDotPlotDataFrame}
\title{Build up dataframe for plotting dot plot with ggplot2}
\usage{
buildDotPlotDataFrame(net.clustering, cluster.names, node.names)
}
\arguments{
\item{net.clustering}{Clustering}

\item{cluster.names}{Cluster names}

\item{node.names}{Node names}
}
\description{
Internal function
}
\keyword{internal}

