% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet.R
\name{cv.fold}
\alias{cv.fold}
\title{Make folds}
\usage{
cv.fold(n, folds = 10)
}
\arguments{
\item{n}{no descr}

\item{folds}{no descr}
}
\value{
no descr
}
\description{
Make folds
}
\author{
n.stadler
}
\keyword{internal}
