% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est2.my.ev2.diffregr}
\alias{est2.my.ev2.diffregr}
\title{Compute weights of sum-w-chi2 (2nd order simplification)}
\usage{
est2.my.ev2.diffregr(y1, y2, x1, x2, beta1, beta2, beta, act1, act2, act)
}
\arguments{
\item{y1}{no descr}

\item{y2}{no descr}

\item{x1}{no descr}

\item{x2}{no descr}

\item{beta1}{no descr}

\item{beta2}{no descr}

\item{beta}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}

\item{act}{no descr}
}
\value{
no descr
}
\description{
*expansion of W in two directions ("dimf>dimg direction" & "dimf>dimg direction")
*simplified computation of weights is obtained by assuming H0 and that X_u~X_v holds
}
\author{
n.stadler
}
\keyword{internal}

