% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est2.ww.mat2}
\alias{est2.ww.mat2}
\title{Weights of sum-w-chi2}
\usage{
est2.ww.mat2(sig1, sig2, sig, act1, act2, act, include.mean = FALSE)
}
\arguments{
\item{sig1}{no descr}

\item{sig2}{no descr}

\item{sig}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}

\item{act}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Compute weights of sum-w-chi2 (1st order simplification)
}
\author{
n.stadler
}
\keyword{internal}

