% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EXPStep.mix}
\alias{EXPStep.mix}
\title{Performs EStep}
\usage{
EXPStep.mix(logphi, mix.prob)
}
\arguments{
\item{logphi}{no descr}

\item{mix.prob}{no descr}
}
\value{
list consiting of
\item{u}{responsibilities}
\item{LL}{loglikelihood}
}
\description{
Performs EStep
}
\author{
n.stadler
}
\keyword{internal}

