% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glasso.parcor}
\alias{glasso.parcor}
\title{Graphical Lasso based on partial correlation penalty}
\usage{
glasso.parcor(s, rho, penalize.diagonal, maxiter = 1000, term = 10^{     -3
  })
}
\arguments{
\item{s}{no descr}

\item{rho}{no descr}

\item{penalize.diagonal}{no descr}

\item{maxiter}{no descr}

\item{term}{no descr}
}
\value{
w; wi; iter
}
\description{
Graphical Lasso based on partial correlation penalty
}
\author{
n.stadler
}
\keyword{internal}

