% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.highdimT2}
\alias{gsea.highdimT2}
\title{GSA based on HighdimT2}
\usage{
gsea.highdimT2(x1, x2, gene.sets, gene.names, gs.names = NULL,
  method = "test.sd", method.p.adjust = "fdr")
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{gene.sets}{no descr}

\item{gene.names}{no descr}

\item{gs.names}{no descr}

\item{method}{no descr}

\item{method.p.adjust}{no descr}
}
\value{
no descr
}
\description{
GSA based on HighdimT2
}
\author{
n.stadler
}
\keyword{internal}

