% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.iriz.scale}
\alias{gsea.iriz.scale}
\title{Irizarry approach (scale only)}
\usage{
gsea.iriz.scale(x1, x2, gene.sets, gene.names, gs.names = NULL,
  method.p.adjust = "fdr", alternative = "two-sided")
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{gene.sets}{no descr}

\item{gene.names}{no descr}

\item{gs.names}{no descr}

\item{method.p.adjust}{no descr}

\item{alternative}{no descr}
}
\value{
no descr
}
\description{
Irizarry approach (scale only)
}
\author{
n.stadler
}
\keyword{internal}

