% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invcov2parcor}
\alias{invcov2parcor}
\title{Convert inverse covariance to partial correlation}
\usage{
invcov2parcor(invcov)
}
\arguments{
\item{invcov}{Inverse covariance matrix}
}
\value{
The partial correlation matrix.
}
\description{
Convert inverse covariance to partial correlation
}
\examples{
inv.cov = generate_inv_cov(p=25)
p.corr = invcov2parcor(inv.cov)
}

