% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logratio.diffregr}
\alias{logratio.diffregr}
\title{Log-likelihood ratio statistics for Differential Regression.}
\usage{
logratio.diffregr(y1, y2, y, xx1, xx2, xx, beta1, beta2, beta)
}
\arguments{
\item{y1}{Response vector condition 1.}

\item{y2}{Response vector condition 2.}

\item{y}{Pooled response vector.}

\item{xx1}{Predictor matrix condition 1.}

\item{xx2}{Predictor matrix condition 2.}

\item{xx}{Pooled predictor matrix}

\item{beta1}{Regression coefficients condition 1.}

\item{beta2}{Regression coefficients condition 2.}

\item{beta}{Pooled regression coefficients.}
}
\value{
2 times log-likelihood ratio statistics.
}
\description{
Log-likelihood ratio statistics for Differential Regression.
}
\author{
n.stadler
}
\keyword{internal}

