% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_cv1se.lasso}
\alias{screen_cv1se.lasso}
\title{Cross-validated Lasso screening (lambda.1se-rule)}
\usage{
screen_cv1se.lasso(x, y)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{Response vector}
}
\value{
Active-set
}
\description{
Cross-validated Lasso screening (lambda.1se-rule)
}
\examples{
screen_cv1se.lasso(matrix(rnorm(5000),50,100),rnorm(50))
}
\author{
n.stadler
}

