% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_cvmin.lasso}
\alias{screen_cvmin.lasso}
\title{Cross-validation lasso screening (lambda.min-rule)}
\usage{
screen_cvmin.lasso(x, y)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{Response vector}
}
\value{
Active-set
}
\description{
Cross-validated Lasso screening (lambda.min-rule)
}
\examples{
screen_cvmin.lasso(matrix(rnorm(5000),50,100),rnorm(50))
}
\author{
n.stadler
}

