% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_mb}
\alias{screen_mb}
\title{Node-wise Lasso-regressions for GGM estimation}
\usage{
screen_mb(x, include.mean = NULL, folds = 10, length.lambda = 20,
  lambdamin.ratio = ifelse(ncol(x) > nrow(x), 0.01, 0.001),
  penalize.diagonal = FALSE, trunc.method = "linear.growth", trunc.k = 5,
  plot.it = FALSE, se = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{The input data. Needs to be a num.samples by dim.samples matrix.}

\item{include.mean}{Include mean in likelihood. TRUE / FALSE (default).}

\item{folds}{Number of folds in the cross-validation (default=10).}

\item{length.lambda}{Length of lambda path to consider (default=20).}

\item{lambdamin.ratio}{Ratio lambda.min/lambda.max.}

\item{penalize.diagonal}{If TRUE apply penalization to diagonal of inverse
covariance as well. (default=FALSE)}

\item{trunc.method}{None / linear.growth (default) / sqrt.growth}

\item{trunc.k}{truncation constant, number of samples per predictor (default=5)}

\item{plot.it}{TRUE / FALSE (default)}

\item{se}{default=FALSE.}

\item{verbose}{If TRUE, output la.min, la.max and la.opt (default=FALSE).}
}
\value{
Returns a list with named elements 'rho.opt', 'wi'.
         Variable rho.opt is the optimal (scaled) penalization parameter (rho.opt=2*la.opt/n).
         The variables wi is a matrix of size dim.samples by dim.samples
         containing the truncated inverse covariance matrix. Variable
         Mu mean of the input data.
}
\description{
Node-wise Lasso-regressions for GGM estimation
}
\details{
(Meinshausen-Buehlmann approach)
}
\examples{
n=50
p=5
x=matrix(rnorm(n*p),n,p)
wihat=screen_mb(x)$wi
}
\author{
n.stadler
}

