% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_mb2}
\alias{screen_mb2}
\title{Screen_mb2}
\usage{
screen_mb2(x, include.mean = NULL, length.lambda = 20,
  trunc.method = "linear.growth", trunc.k = 5, plot.it = FALSE,
  verbose = FALSE)
}
\arguments{
\item{x}{no descr}

\item{include.mean}{no descr}

\item{length.lambda}{no descr}

\item{trunc.method}{no descr}

\item{trunc.k}{no descr}

\item{plot.it}{no descr}

\item{verbose}{no descr}
}
\value{
no descr
}
\description{
Screen_mb2
}
\author{
n.stadler
}
\keyword{internal}

