% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.diffnet}
\alias{summary.diffnet}
\title{Summary function for object of class 'diffnet'}
\usage{
summary.diffnet(x)
}
\arguments{
\item{x}{object of class 'diffnet'}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'diffnet'
}
\author{
nicolas
}

