% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ggmgsa}
\alias{summary.ggmgsa}
\title{Summary function for object of class 'ggmgsa'}
\usage{
summary.ggmgsa(x)
}
\arguments{
\item{x}{object of class 'ggmgsa'}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'ggmgsa'
}
\author{
nicolas
}

