% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.nethetclustering}
\alias{summary.nethetclustering}
\title{Summary function for object of class 'nethetclustering'}
\usage{
summary.nethetclustering(x)
}
\arguments{
\item{x}{object of class 'nethetclustering'}
}
\value{
Network statistics (a 'nethetsummary' object)
}
\description{
Summary function for object of class 'nethetclustering'
}
\author{
frankd
}

