% Generated by roxygen2 (4.0.2): do not edit by hand
\name{symmkldist}
\alias{symmkldist}
\title{Compute symmetric kull-back leibler distance}
\usage{
symmkldist(mu1, mu2, sig1, sig2)
}
\arguments{
\item{mu1}{no descr}

\item{mu2}{no descr}

\item{sig1}{no descr}

\item{sig2}{no descr}
}
\value{
symmetric kull-back leibler distance
}
\description{
Compute symmetric kull-back leibler distance
}
\author{
n.stadler
}
\keyword{internal}

