% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test.sd}
\alias{test.sd}
\title{High-Dim Two-Sample Test (Srivastava, 2006)}
\usage{
test.sd(x1, x2)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}
}
\value{
no descr
}
\description{
High-Dim Two-Sample Test (Srivastava, 2006)
}
\author{
n.stadler
}
\keyword{internal}

