% Generated by roxygen2 (4.0.2): do not edit by hand
\name{w.kldist}
\alias{w.kldist}
\title{Distance between comps based on symm. kl-distance}
\usage{
w.kldist(Mu, Sig)
}
\arguments{
\item{Mu}{no descr}

\item{Sig}{no descr}
}
\value{
list consisting of
\item{comp.kldist}{}
\item{min.comp.kldist}{}
}
\description{
Distance between comps based on symm. kl-distance
}
\author{
n.stadler
}
\keyword{internal}

