% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ww.mat2.diffregr}
\alias{ww.mat2.diffregr}
\title{Computation M matrix and eigenvalues}
\usage{
ww.mat2.diffregr(Sig, act, act1, act2)
}
\arguments{
\item{Sig}{no descr}

\item{act}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}
}
\value{
no descr
}
\description{
Computation M matrix and eigenvalues
}
\author{
n.stadler
}
\keyword{internal}

